// lib/api/generated/models/planWeek.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CKS Slavia API
 * API panelu klubowego CKS Slavia Ruda Śląska
 * OpenAPI spec version: 2.2.0.1+1
 */
import type { PlanDay } from './planDay';

export interface PlanWeek {
  days?: PlanDay[];
  id: string;
  /** @minimum 0 */
  index: number;
}